# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FrozenBot API
  description: |
    API for FrozenBot - Telegram bot for frozen food delivery system.

    This API serves three main clients:
    - **Telegram Bot**: For user interactions and order processing
    - **Admin Panel**: React-based admin interface for managing orders, products, and analytics
    - **WebApp**: Telegram WebApp for customer ordering interface

    ## Authentication

    The API uses multiple authentication methods:
    - **JWT Tokens**: For admin panel authentication
    - **Telegram WebApp Validation**: For WebApp users using initData
    - **Bot Token**: For internal bot operations

    ## Rate Limiting

    API endpoints are rate limited to prevent abuse:
    - Admin endpoints: 100 requests per minute
    - User endpoints: 60 requests per minute
    - Public endpoints: 30 requests per minute

  version: 1.0.0
  contact:
    name: FrozenBot API Support
    email: support@frozenbot.com
  license:
    name: MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.frozenbot.com
    description: Production server
...
paths:
...
  /api/payments/webhook:
    post:
      tags: [payments]
      summary: Payment webhook
      description: Handle payment provider webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentWebhookRequest'
      responses:
        '200':
          description: Webhook processed successfully
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/payments/{order_id}/status:
    get:
      tags: [payments]
      summary: Get payment status
      description: Get payment status for order
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
          description: Order ID
      security:
        - JWTAuth: []
        - TelegramWebApp: []
      responses:
        '200':
          description: Payment status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    JWTAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication

    TelegramWebApp:
      type: apiKey
      in: header
      name: X-Telegram-Init-Data
      description: Telegram WebApp initData for user authentication
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...
    PaymentWebhookRequest:
      type: object
      required:
        - order_id
        - status
        - amount
      properties:
        order_id:
          type: integer
          example: 1
        status:
          type: string
          enum:
            - success
            - failed
            - pending
          example: "success"
        amount:
          type: number
          format: float
          example: 450.0
        transaction_id:
          type: string
          example: "txn_123456789"
        payment_method:
          type: string
          example: "card"
        metadata:
          type: object
          additionalProperties: true
    PaymentStatus:
      type: object
      required:
        - order_id
        - status
        - amount
      properties:
        order_id:
          type: integer
          example: 1
        status:
          type: string
          enum:
            - pending
            - success
            - failed
            - refunded
          example: "success"
        amount:
          type: number
          format: float
          example: 450.0
        transaction_id:
          type: string
          nullable: true
          example: "txn_123456789"
        payment_method:
          type: string
          example: "card"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "bad_request"
            message: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
            message: "Authentication required"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "not_found"
            message: "Resource not found"
...